refactor: BonusNumber 이동에 따른 commit
특별히 변경된 코드는 없지만 commit을 다시 하라길래 add 후 커밋
