+++ src/main/java/lotto/view/InputView.java
@@ -2,9 +2,6 @@ package lotto.view;
 
 import static camp.nextstep.edu.missionutils.Console.readLine;
 
-import java.io.Console;
-import lotto.util.Information;
-
 public class InputView {
 
     public static String readInput(String message) {
